Create field.ts
