+++ src/utility.ts
@@ -1,6 +1,4 @@
-import { Attribute, FieldAttribute } from "./attribute";
-import { logWarning } from "./log";
-import { isKeyword } from "./lua";
+import { Attribute } from "./attribute";
 
 export function stripGenericParams(text: string) {
   const index = text.indexOf("<");
@@ -29,28 +27,6 @@ export function joinLines(dest: string, src: string) {
 export function formatAttribute(attribute: Readonly<Attribute>): string {
   const { type, rawText: description } = attribute;
   return `@${type}${description}`;
-}
-
-export function generateField(rule: FieldAttribute, indent: string): string {
-  if (rule.field == null) {
-    logWarning(
-      `Invalid attribute, field name expected: ${formatAttribute(rule)}`
-    );
-  }
-  if (rule.field == null) {
-    console.error(`No field property`, rule);
-    return "";
-  }
-  var { name, description } = rule.field;
-  return formatField(name, description.trimStart(), indent);
-}
-
-export function formatField(name: string, description: string, indent: string) {
-  const fieldName = isKeyword(name) ? `["${name}"]` : name;
-  return (
-    toLuaComment(`@type ${description.trimEnd()}`, indent) +
-    `\n${indent}${fieldName} = nil`
-  );
 }
 
 export function toLuaComment(text: string, indent = ""): string {
